MOD: remove network grid from winds
Removing the need for the user to provide a network grid. Instead, in
(private) functions like the echo bounds and fall speed, all grids are
used to determine the maximum observed value (reflectivity) at each grid
point.

Furthermore, this commit enhances the output grid of the wind retrieval.
Now the grid (radar) with the latest volume scan and consequently the
latest end time is used for the axes dictionary. The user is also given
the option of saving the reflectivity field of the network, which is
simply the maximum observed reflectivity from all grids.