ENH: wind retrieval grid object
Several enhancements being made to the output grid object of
solve_wind_field. This mainly focuses on the time axes of the grid,
which now includes the ARM standard base_time and time_offset variables.
Furthermore, the volume start and end times now correspond to the
earliest volume start time and the latest volume end time, respectively.
This provides the user with valuable information about the full time window
the retrieval may correspond to. Finally, the time axis now corresponds
to the time half-way between the start and end times.