BUG: normalized divergence calculation
Fixing bug in the normalized divergence calculation. I did not include
one of the terms from the anelastic continuity equation in the
calculation. What this meant was that I was looking at how well the wind
retrieval satisfied an incompressible atmosphere, not an anelastic
atmosphere. Unfortunately the test case that I ran with this bug fixed
still did not produce "optimal" values of the normalized divergence...