Additional data masking criteria
I am adding the ability to mask grid points that are close to the radar,
where close is defined using a specified radius of influence. Basically
radial velocity observations close to a radar typically produce spurious
air motion retrievals close to said radar due to (I think) the nature of
the geometry close to the radar.